moved component specific constants to that component and fixed imports in utils file
