+++ src/Components/QuickTimesDropdown/QuickTimesDropdown.stories.tsx
@@ -2,9 +2,11 @@ import React from 'react';
 import { ComponentStory } from '@storybook/react';
 import { getDefaultStoryDecorator } from '../../Models/Services/StoryUtilities';
 import QuickTimesDropdown from './QuickTimesDropdown';
-import { IQuickTimesDropdownProps } from './QuickTimesDropdown.types';
+import {
+    IQuickTimesDropdownProps,
+    QuickTimeSpanKey
+} from './QuickTimesDropdown.types';
 import { Icon, IDropdownOption } from '@fluentui/react';
-import { QuickTimeSpanKey } from '../../Models/Constants/Enums';
 
 const wrapperStyle = { width: '200px', height: '600px', padding: 8 };
