+++ src/Models/Constants/Constants.ts
@@ -2,8 +2,7 @@ import {
     AzureAccessPermissionRoleGroups,
     AzureAccessPermissionRoles,
     IADTBackgroundColor,
-    IADTObjectColor,
-    QuickTimeSpanKey
+    IADTObjectColor
 } from '../Constants';
 import {
     defaultDataHistoryWidget,
@@ -543,20 +542,3 @@ export const RequiredAccessRoleGroupForADTInstance: AzureAccessPermissionRoleGro
 };
 
 export const CONNECTION_STRING_SUFFIX = '.kusto.windows.net';
-
-/** Quick time span key to value in millisecond mapping */
-export const QuickTimeSpans = {
-    [QuickTimeSpanKey.Last15Mins]: 15 * 60 * 1000,
-    [QuickTimeSpanKey.Last30Mins]: 30 * 60 * 1000,
-    [QuickTimeSpanKey.LastHour]: 1 * 60 * 60 * 1000,
-    [QuickTimeSpanKey.Last3Hours]: 3 * 60 * 60 * 1000,
-    [QuickTimeSpanKey.Last6Hours]: 6 * 60 * 60 * 1000,
-    [QuickTimeSpanKey.Last12Hours]: 12 * 60 * 60 * 1000,
-    [QuickTimeSpanKey.Last24Hours]: 24 * 60 * 60 * 1000,
-    [QuickTimeSpanKey.Last7Days]: 7 * 24 * 60 * 60 * 1000,
-    [QuickTimeSpanKey.Last30Days]: 30 * 24 * 60 * 60 * 1000,
-    [QuickTimeSpanKey.Last60Days]: 60 * 24 * 60 * 60 * 1000,
-    [QuickTimeSpanKey.Last90Days]: 90 * 24 * 60 * 60 * 1000,
-    [QuickTimeSpanKey.Last180Days]: 180 * 24 * 60 * 60 * 1000,
-    [QuickTimeSpanKey.LastYear]: 365 * 24 * 60 * 60 * 1000
-};
